Code changes for multiple mpp

+++ RequestCMR/src/com/ibm/cio/cmr/request/automation/util/AutomationUtil.java
@@ -884,7 +884,7 @@ public abstract class AutomationUtil {
       query.append(" and COUNTY= :COUNTY");
       query.setParameter("COUNTY", addrToCheck.getCounty());
     }
-    if (payGoAddredited && "ZP01".equals(addrToCheck.getId().getAddrType())) {
+    if (payGoAddredited) {
       if (addrToCheck.getCustNm4() != null) {
         query.append(" and CUST_NM4 = :CUST_NM4");
         query.setParameter("CUST_NM4", addrToCheck.getCustNm4());
@@ -930,31 +930,17 @@ public abstract class AutomationUtil {
         }
       } else {
         if (!"ZS01".equals(addr.getId().getAddrType())) {
-          if (!"ZP01".equals(addr.getId().getAddrType())) {
-            if (compareCustomerNames(zs01, addr)
-                && (StringUtils.isNotBlank(addr.getAddrTxt()) && addr.getAddrTxt().trim().toUpperCase().equals(mainStreetAddress1))
-                && addr.getCity1().trim().toUpperCase().equals(mainCity) && addr.getPostCd().trim().equals(mainPostalCd)) {
-              details.append("Removing duplicate address record: " + addr.getId().getAddrType() + " from the request.").append("\n");
-              Addr merged = entityManager.merge(addr);
-              if (merged != null) {
-                entityManager.remove(merged);
-              }
-              it.remove();
-              removed = true;
+          if (compareCustomerNames(zs01, addr)
+              && (StringUtils.isNotBlank(addr.getAddrTxt()) && addr.getAddrTxt().trim().toUpperCase().equals(mainStreetAddress1))
+              && addr.getCity1().trim().toUpperCase().equals(mainCity) && addr.getPostCd().trim().equals(mainPostalCd)
+              && addr.getCustNm4().trim().toUpperCase().equals(mainCustNm4)) {
+            details.append("Removing duplicate address record: " + addr.getId().getAddrType() + " from the request.").append("\n");
+            Addr merged = entityManager.merge(addr);
+            if (merged != null) {
+              entityManager.remove(merged);
             }
-          } else if ("ZP01".equals(addr.getId().getAddrType())) {
-            if (compareCustomerNames(zs01, addr)
-                && (StringUtils.isNotBlank(addr.getAddrTxt()) && addr.getAddrTxt().trim().toUpperCase().equals(mainStreetAddress1))
-                && addr.getCity1().trim().toUpperCase().equals(mainCity) && addr.getPostCd().trim().equals(mainPostalCd)
-                && addr.getCustNm4().trim().toUpperCase().equals(mainCustNm4)) {
-              details.append("Removing duplicate address record: " + addr.getId().getAddrType() + " from the request.").append("\n");
-              Addr merged = entityManager.merge(addr);
-              if (merged != null) {
-                entityManager.remove(merged);
-              }
-              it.remove();
-              removed = true;
-            }
+            it.remove();
+            removed = true;
           }
         }
       }
